+++ eslint/index.js
@@ -813,43 +813,6 @@ module.exports = {
 		'import/unambiguous': 'off',
 
 		/**
-		 * eslint-plugin-jsx-a11y
-		 *
-		 * @see https://github.com/jsx-eslint/eslint-plugin-jsx-a11y
-		 */
-		'jsx-a11y/alt-text': 'error',
-		'jsx-a11y/anchor-has-content': 'error',
-		'jsx-a11y/anchor-is-valid': 'error',
-		'jsx-a11y/aria-activedescendant-has-tabindex': 'error',
-		'jsx-a11y/aria-props': 'error',
-		'jsx-a11y/aria-proptypes': 'error',
-		'jsx-a11y/aria-role': 'error',
-		'jsx-a11y/aria-unsupported-elements': 'error',
-		'jsx-a11y/autocomplete-valid': 'error',
-		'jsx-a11y/click-events-have-key-events': 'error',
-		'jsx-a11y/heading-has-content': 'error',
-		'jsx-a11y/html-has-lang': 'off', // <html> elements set by react-helmet automatically have a lang attribute set
-		'jsx-a11y/iframe-has-title': 'error',
-		'jsx-a11y/img-redundant-alt': 'error',
-		'jsx-a11y/interactive-supports-focus': 'error',
-		'jsx-a11y/label-has-associated-control': 'error',
-		'jsx-a11y/media-has-caption': 'error',
-		'jsx-a11y/mouse-events-have-key-events': 'error',
-		'jsx-a11y/no-access-key': 'off', // If you explicitly specify an access key, you usually have a reason for this, so it should not be prevented by a rule
-		'jsx-a11y/no-autofocus': ['error', { ignoreNonDOM: true }],
-		'jsx-a11y/no-distracting-elements': 'error',
-		'jsx-a11y/no-interactive-element-to-noninteractive-role': 'error',
-		'jsx-a11y/no-noninteractive-element-interactions': 'error',
-		'jsx-a11y/no-noninteractive-element-to-interactive-role': 'error',
-		'jsx-a11y/no-noninteractive-tabindex': 'error',
-		'jsx-a11y/no-redundant-roles': 'error',
-		'jsx-a11y/no-static-element-interactions': 'error',
-		'jsx-a11y/role-has-required-aria-props': 'error',
-		'jsx-a11y/role-supports-aria-props': 'error',
-		'jsx-a11y/scope': 'error',
-		'jsx-a11y/tabindex-no-positive': 'error',
-
-		/**
 		 * eslint-plugin-eslint-comments
 		 *
 		 * @see https://mysticatea.github.io/eslint-plugin-eslint-comments/