+++ eslint/overrides-react.js
@@ -14,16 +14,7 @@ module.exports = {
 				 *
 				 * @see https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules
 				 */
-				'@typescript-eslint/member-ordering': 'off', // For React components we are using react/sort-comp
-
-				/**
-				 * eslint-plugin-react-hooks
-				 *
-				 * @see https://github.com/facebook/react/tree/master/packages/eslint-plugin-react-hooks
-				 * @see https://reactjs.org/docs/hooks-rules.html
-				 */
-				'react-hooks/rules-of-hooks': 'error',
-				'react-hooks/exhaustive-deps': 'error'
+				'@typescript-eslint/member-ordering': 'off' // For React components we are using react/sort-comp
 			}
 		}
 	]