+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-### Changed
+### Removed
 
-**Right now, a couple of ESLint plugins (e.g. eslint-plugin-jsx-a11y, eslint-plugin-promise) are still not compatible with ESLint v8. They prevent us from updating.**
+- [eslint] `eslint-plugin-promise` has be removed, because it seems not to be regularly maintained anymore, which blocks us since one month from updating to ESLint v8.
 
 ### Changed